+++ shapemap.h
@@ -440,7 +440,7 @@ class ShapeMap : public PixelBase {
 
   public:
     const std::string &getName() const { return m_name; }
-    int addAttribute(const std::string &name) { return m_attributes->insertOrResetColumn(name); }
+    size_t addAttribute(const std::string &name) { return m_attributes->insertOrResetColumn(name); }
     void removeAttribute(int col) { m_attributes->removeColumn(col); }
     // I don't want to do this, but every so often you will need to update this table
     // use const version by preference
